requirements.txt
Django==5.0.0
djangorestframework==3.14.0
mysqlclient==2.2.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
gunicorn>=20.1.0
django-cors-headers
celery
redis